ui changes  - update status bar style  - add search ui to holdings page
