+++ screens/home/index.tsx
@@ -6,16 +6,15 @@ import { adjustFontSizeForIOS, formatAmountWithWholeAndDecimal } from '@/utils/h
 import { LQDButton, LQDPoolPairCard, LQDPoolPairPaper, SearchUI } from '@/components';
 import { CaretRightIcon, DirectUpIcon, DollarSquareIcon, SearchIcon, SettingsIcon, TrendUpIcon } from '@/assets/icons';
 import { useAccountActions } from '@/store/account/actions';
-import useAppActions from '@/store/app/actions';
 import { usePoolActions } from '@/store/pools/actions';
 import { useOnMount } from '@/hooks/useOnMount';
 import Section from './section';
+import SearchPlaceholder from '@/components/search-ui/search-placeholder';
 
 const Home = () => {
   const { router, poolsState, smartAccountState, accountState, appState } = useSystemFunctions();
   const { getTokens } = useAccountActions();
   const { getPools, getAllPools } = usePoolActions();
-  const { searchIsFocused: focusSearch, showSearch } = useAppActions();
 
   const { trendingPools, hotPools, topGainers } = poolsState;
 
@@ -28,11 +27,6 @@ const Home = () => {
   const top10TrendingPools = trendingPoolsArray?.slice(0, 10);
   const top10HotPools = hotPoolsArray.slice(0, 10) ?? [];
   const top7Gainers = top7GainersArray.slice(0, 7);
-
-  const focusInput = () => {
-    focusSearch(true);
-    showSearch(true);
-  };
 
   const sections = [
     {
@@ -104,21 +98,7 @@ const Home = () => {
 
   return (
     <>
-      <View style={styles.searchInnerWrapper}>
-        <View style={styles.inputWrapper}>
-          <TouchableOpacity style={styles.inputContainer} onPress={focusInput}>
-            <SearchIcon />
-
-            <View>
-              <Text style={styles.inputPlaceholderText}>Search...</Text>
-            </View>
-          </TouchableOpacity>
-        </View>
-
-        <TouchableOpacity onPress={() => router.navigate('/(settings)')}>
-          <SettingsIcon />
-        </TouchableOpacity>
-      </View>
+      <SearchPlaceholder />
 
       <ScrollView style={styles.container} contentContainerStyle={styles.contentContainer} showsVerticalScrollIndicator={false}>
         <View style={styles.balanceAndActionContainer}>
@@ -226,46 +206,5 @@ const styles = StyleSheet.create({
     backgroundColor: '#fff',
     paddingTop: Platform.OS === 'android' ? RNStatusBar.currentHeight : 20,
     paddingBottom: Platform.OS === 'android' ? -(RNStatusBar.currentHeight || 0) : -48,
-  },
-
-  searchInnerWrapper: {
-    paddingHorizontal: 16,
-    backgroundColor: '#fff',
-    paddingTop: Platform.OS === 'android' ? RNStatusBar.currentHeight : 20,
-    paddingBottom: Platform.OS === 'android' ? -(RNStatusBar.currentHeight || 0) : -48,
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-
-  inputWrapper: {
-    paddingHorizontal: 16,
-    paddingVertical: 10,
-    backgroundColor: '#fff',
-    width: '90%',
-  },
-
-  inputContainer: {
-    paddingHorizontal: 16,
-    paddingVertical: 10,
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-    borderRadius: 13,
-    borderColor: '#EAEEF4',
-    borderWidth: 1,
-    shadowColor: 'rgba(15, 23, 42, 0.04)',
-    shadowOffset: { width: 0, height: 1 },
-    shadowOpacity: 0.3,
-    shadowRadius: 3,
-    elevation: 1,
-    backgroundColor: '#fff',
-    width: '100%',
-  },
-
-  inputPlaceholderText: {
-    fontSize: adjustFontSizeForIOS(14, 2),
-    lineHeight: 18.48,
-    color: '#94A3B8',
-    fontFamily: 'AeonikRegular',
   },
 });
